videosummarizer/requirements.txt
moviepy
openai-whisper
transformers
torch